Clean Up and Updated URL
